refactor(xregexp): extract regexp literal extraction into a package
The required-literal walk was duplicated: chsql.RegexpLiterals for the
ClickHouse hasToken prefilters and storagebackend.regexpLiterals for the
bloom token hints, identical but for the return type. It is also not SQL, so
chsql was the wrong home for it.

Move it to internal/xregexp.Literals, alongside the other stdlib extension
packages, and point both backends at it.
