+++ internal/xregexp/xregexp_test.go
@@ -1,4 +1,4 @@
-package chsql
+package xregexp
 
 import (
 	"fmt"
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestRegexpLiterals(t *testing.T) {
+func TestLiterals(t *testing.T) {
 	tests := []struct {
 		pattern string
 		want    []string
@@ -29,15 +29,15 @@ func TestRegexpLiterals(t *testing.T) {
 	}
 	for i, tt := range tests {
 		t.Run(fmt.Sprintf("Test%d", i+1), func(t *testing.T) {
-			require.Equal(t, tt.want, RegexpLiterals(tt.pattern))
+			require.Equal(t, tt.want, Literals(tt.pattern))
 		})
 	}
 }
 
-// FuzzRegexpLiterals asserts the required-literal invariant: any string that matches the pattern
-// contains every extracted literal as a substring — otherwise a hasToken prefilter built from it
-// would prune a genuine match.
-func FuzzRegexpLiterals(f *testing.F) {
+// FuzzLiterals asserts the required-literal invariant: any string matching the pattern contains
+// every extracted literal as a substring — otherwise a prefilter built from it would prune a
+// genuine match.
+func FuzzLiterals(f *testing.F) {
 	for _, p := range []string{`foo.*bar`, `error.*timeout`, `foo(bar)?baz`, `^user logged in$`} {
 		f.Add(p, "xfoobarx")
 	}
@@ -46,7 +46,7 @@ func FuzzRegexpLiterals(f *testing.F) {
 		if err != nil || !re.MatchString(value) {
 			return // only matching values constrain the invariant
 		}
-		for _, lit := range RegexpLiterals(pattern) {
+		for _, lit := range Literals(pattern) {
 			require.Containsf(t, value, lit,
 				"value %q matches %q but lacks required literal %q", value, pattern, lit)
 		}
